+++ pkg/scaling/scalers_builder.go
@@ -192,6 +192,8 @@ func buildScaler(ctx context.Context, client client.Client, triggerType string, 
 		return scalers.NewGcpCloudTasksScaler(config)
 	case "gcp-pubsub":
 		return scalers.NewPubSubScaler(config)
+	case "gcp-spanner":
+		return scalers.NewGcpSpannerScaler(config)
 	case "gcp-stackdriver":
 		return scalers.NewStackdriverScaler(ctx, config)
 	case "gcp-storage":